pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis-satellite"
version = "0.1.0"
description = "A HomeAssistant Wyoming Satellite. Runs on RPI4 with a DigiAmp+ and Respeaker XMOS3800 mic array"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "wyoming-satellite>=1.5.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0"
]

[tool.setuptools]
packages = ["jarvis_satellite"]

